requirements.txt
opencv-python
numpy
mss
Pillow
pyautogui
pytesseract
loguru